refactored monitor handler interaction to better allow differing event types in same system

+++ core/functionality.py
@@ -14,28 +14,12 @@ from papermill.translators import papermill_translators
 from typing import Any, Union
 from random import SystemRandom
 
-from core.meow import BasePattern, BaseRecipe, BaseRule
+#from core.meow import BasePattern, BaseRecipe, BaseRule
 from core.correctness.validation import check_type, valid_dict, valid_list, \
     valid_existing_file_path, valid_path
 from core.correctness.vars import CHAR_LOWERCASE, CHAR_UPPERCASE, \
     VALID_CHANNELS, HASH_BUFFER_SIZE, SHA256, DEBUG_WARNING, DEBUG_ERROR, \
-    DEBUG_INFO
-
-def check_pattern_dict(patterns, min_length=1):
-    valid_dict(patterns, str, BasePattern, strict=False, min_length=min_length)
-    for k, v in patterns.items():
-        if k != v.name:
-            raise KeyError(f"Key '{k}' indexes unexpected Pattern '{v.name}' "
-                "Pattern dictionaries must be keyed with the name of the "
-                "Pattern.")
-
-def check_recipe_dict(recipes, min_length=1):
-    valid_dict(recipes, str, BaseRecipe, strict=False, min_length=min_length)
-    for k, v in recipes.items():
-        if k != v.name:
-            raise KeyError(f"Key '{k}' indexes unexpected Recipe '{v.name}' "
-                "Recipe dictionaries must be keyed with the name of the "
-                "Recipe.")
+    DEBUG_INFO, EVENT_TYPE
 
 def generate_id(prefix:str="", length:int=16, existing_ids:list[str]=[], 
         charset:str=CHAR_UPPERCASE+CHAR_LOWERCASE, attempts:int=24):
@@ -47,45 +31,6 @@ def generate_id(prefix:str="", length:int=16, existing_ids:list[str]=[],
             return id
     raise ValueError(f"Could not generate ID unique from '{existing_ids}' "
         f"using values '{charset}' and length of '{length}'.")
-
-def create_rules(patterns:Union[dict[str,BasePattern],list[BasePattern]], 
-        recipes:Union[dict[str,BaseRecipe],list[BaseRecipe]],
-        new_rules:list[BaseRule]=[])->dict[str,BaseRule]:
-    check_type(patterns, dict, alt_types=[list])
-    check_type(recipes, dict, alt_types=[list])
-    valid_list(new_rules, BaseRule, min_length=0)
-
-    if isinstance(patterns, list):
-        valid_list(patterns, BasePattern, min_length=0)
-        patterns = {pattern.name:pattern for pattern in patterns}
-    else:
-        check_pattern_dict(patterns, min_length=0)
-
-    if isinstance(recipes, list):
-        valid_list(recipes, BaseRecipe, min_length=0)
-        recipes = {recipe.name:recipe for recipe in recipes}
-    else:
-        check_recipe_dict(recipes, min_length=0)
-
-    # Imported here to avoid circular imports at top of file
-    import rules
-    rules = {}
-    all_rules ={(r.pattern_type, r.recipe_type):r for r in [r[1] \
-        for r in inspect.getmembers(sys.modules["rules"], inspect.isclass) \
-        if (issubclass(r[1], BaseRule))]}
-
-    for pattern in patterns.values():
-        if pattern.recipe in recipes:
-            key = (type(pattern).__name__, 
-                type(recipes[pattern.recipe]).__name__)
-            if (key) in all_rules:
-                rule = all_rules[key](
-                    generate_id(prefix="Rule_"), 
-                    pattern, 
-                    recipes[pattern.recipe]
-                )
-                rules[rule.name] = rule
-    return rules
 
 def wait(inputs:list[VALID_CHANNELS])->list[VALID_CHANNELS]:
     all_connections = [i for i in inputs if type(i) is Connection] \
@@ -133,6 +78,8 @@ def rmtree(directory:str):
 
     :return: No return
     """
+    if not os.path.exists(directory):
+        return
     for root, dirs, files in os.walk(directory, topdown=False):
         for file in files:
             os.remove(os.path.join(root, file))
@@ -296,3 +243,7 @@ def print_debug(print_target, debug_level, msg, level)->None:
                 elif level == DEBUG_WARNING:
                     status = "WARNING"
                 print(f"{status}: {msg}", file=print_target)
+
+def create_event(event_type:str, source:dict[Any,Any]={})->dict[Any,Any]:
+    return {**source, EVENT_TYPE: event_type}
+
